feat(component): add ContentBlockSimple – replaces SimpleLongForm
