+++ packages/react/src/patterns/blocks/ContentBlockSimple/README.md
