+++ packages/react/src/patterns/blocks/ContentBlockSimple/__stories__/index.scss
@@ -1,0 +1,1 @@
+@import '../../../../../../styles/scss/patterns/blocks/content-block-simple/content-block-simple';
