--- packages/react/src/patterns/blocks/SimpleLongForm/__stories__/index.scss
@@ -1,1 +1,0 @@
-@import '../../../../../../styles/scss/patterns/blocks/simplelongform/simplelongform';
