+++ packages/react/src/patterns/blocks/index.js
@@ -6,9 +6,9 @@
  */
 
 export * from './ContentBlockMedia';
+export * from './ContentBlockSimple';
 export * from './ContentGroupCards';
 export * from './FeaturedLink';
 export * from './ListSection';
 export * from './LogoGrid';
 export * from './PictogramArray';
-export * from './SimpleLongForm';
