--- packages/styles/scss/patterns/blocks/simplelongform/README.md
@@ -1,38 +1,0 @@
-### SCSS
-
-#### Usage
-
-Import the package css into the top of your main CSS file.
-
-```css
-@import '@carbon/ibmdotcom-styles/scss/patterns/blocks/simplelongform/index';
-```
-
-#### Modifiers
-
-Use these modifiers with `.bx--simplelongform` class.
-
-| Class                              | Description                                           |
-| ---------------------------------- | ----------------------------------------------------- |
-| `.bx--simplelongform--icon-link`   | The className for simple long form with icon link     |
-| `.bx--simplelongform--card`        | The className for simple long form with card link     |
-| `.bx--simplelongform--with-border` | The className for simple long form with border bottom |
-
-#### Elements
-
-| Class                              | Description                                                                     |
-| ---------------------------------- | ------------------------------------------------------------------------------- |
-| `.bx--simplelongform__container`   | Element containing all simple long form content                                 |
-| `.bx--simplelongform__row`         | Vertical divisions across the viewport                                          |
-| `.bx--simplelongform__col`         | Horizonal divisions across the viewport and change in width at different screen |
-| `.bx--simplelongform__title`       | Title for simple long form pattern                                              |
-| `.bx--simplelongform__content`     | Content holder for short description                                            |
-| `.bx--simplelongform__link`        | Wrapper for link component                                                      |
-| `.bx--simplelongform__divider_col` | Horizonal division for divider element                                          |
-| `.bx--simplelongform__divider`     | Divider element                                                                 |
-
-#### Blocks
-
-| Class                 | Description                                               |
-| --------------------- | --------------------------------------------------------- |
-| `.bx--simplelongform` | Highest level class name for the simple long form pattern |
